Major edits to the BST.
BST is not a child class of the Binary_Tree anymore. The variables and 
methods would conflict a lot. Also, there were some methods of the 
Binary_Tree that would should be prohibited in the BST class. Therefore, 
it was more accurate to remove the parent-child relationship.

BST has its own functions now. The definition and implementation of the 
trees were separated between trees.hpp and trees.cpp files.

The demo.cpp file was extended to show more functionalities of the 
library. The BST is almost completed. The deletion method has still to 
be implemented.